# docs/openapi.json
{
	"openapi": "3.1.0",
	"info": { "title": "SocialAnimal REST", "version": "1.0.0" },
	"components": {
		"securitySchemes": {
			"bearerAuth": { "type": "http", "scheme": "bearer" }
		}
	},
	"paths": {
		"/api/v1/auth/registrations": {
			"post": {
				"operationId": "register",
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"additionalProperties": false,
								"properties": {
									"email": {
										"type": "string",
										"format": "email",
										"maxLength": 254
									},
									"password": {
										"type": "string",
										"minLength": 8,
										"maxLength": 72
									},
									"name": {
										"type": "string",
										"maxLength": 100
									},
									"inviteCode": { "type": "string" }
								},
								"required": ["email", "password"]
							}
						}
					}
				},
				"responses": {
					"201": {
						"description": "201",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"id": { "type": "string" },
										"email": { "type": "string" },
										"name": { "type": ["null", "string"] },
										"isAdmin": { "type": "boolean" },
										"createdAt": {
											"type": "string",
											"format": "date-time"
										}
									},
									"required": [
										"id",
										"email",
										"name",
										"isAdmin"
									]
								}
							}
						}
					},
					"400": {
						"description": "400",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					},
					"401": {
						"description": "401",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					},
					"403": {
						"description": "403",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					},
					"404": {
						"description": "404",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					},
					"409": {
						"description": "409",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					},
					"422": {
						"description": "422",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					},
					"500": {
						"description": "500",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					},
					"502": {
						"description": "502",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					},
					"503": {
						"description": "503",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					}
				}
			}
		},
		"/api/v1/auth/sessions": {
			"post": {
				"operationId": "login",
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"additionalProperties": false,
								"properties": {
									"email": {
										"type": "string",
										"format": "email",
										"maxLength": 254
									},
									"password": {
										"type": "string",
										"minLength": 1,
										"maxLength": 72
									}
								},
								"required": ["email", "password"]
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "200",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"token": { "type": "string" },
										"user": {
											"type": "object",
											"additionalProperties": false,
											"properties": {
												"id": { "type": "string" },
												"email": { "type": "string" },
												"name": {
													"type": ["null", "string"]
												},
												"isAdmin": {
													"type": "boolean"
												},
												"createdAt": {
													"type": "string",
													"format": "date-time"
												}
											},
											"required": [
												"id",
												"email",
												"name",
												"isAdmin"
											]
										}
									},
									"required": ["token", "user"]
								}
							}
						}
					},
					"400": {
						"description": "400",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					},
					"401": {
						"description": "401",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					},
					"403": {
						"description": "403",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					},
					"404": {
						"description": "404",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					},
					"409": {
						"description": "409",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					},
					"422": {
						"description": "422",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					},
					"500": {
						"description": "500",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					},
					"502": {
						"description": "502",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					},
					"503": {
						"description": "503",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					}
				}
			}
		},
...
	}
}